backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.8.1
chromadb==0.6.3
openai==1.75.0
anthropic==0.50.0
sentence-transformers==4.1.0
PyMuPDF==1.25.5
pdfplumber==0.11.6
httpx==0.28.1
orjson==3.10.16
tenacity==9.1.2
cachetools==5.5.2
sqlalchemy==2.0.40
asyncpg==0.30.0
aiosqlite==0.21.0
greenlet==3.2.1
alembic==1.15.2
redis==5.2.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
aioboto3==13.4.0
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357
gunicorn==23.0.0
sentry-sdk[fastapi]==2.27.0
